refactor(client): Provided go client.AddResource([]byte, string, pb.WorkflowRequestObject_ResourceType) method
Provided a direct way to add a byte array with name and type to a command workflow,
besides client.AddResourceFile(string) limited to filesystem. This could be useful
for the case when *.bpmn descriptor is bundled within executable binary, with
popular go-bindata asset generator and possibly some others, including runtime
acquisition in read-only fs environment.