fix(clients/zbctl): compare the broker index with the broker amount
Previously the zbctl output only worked when the partition and broker count were equal.
This was because a line break was only done as long as the index for iterating Brokers
or the index for iterating partitions was smaller than the amount of partitions -1.

This commit fixes by comparing against the broker amount.