fix: print correct json input
Previously on test failure, the input was overwritten by the return of the reformatJson function, which gave not much helpful error message. The return value is stored in a intermediate variable to be overwrite the real variable only in the case of success, such that we print the correct input.