test(go): print out container logs per failed test case
At the moment, the container suite works by printing out the container
logs when a test fails, at the suite level. This means, for zbctl's
tests, when `TestCommonCommands` fails. However, `TestCommonCommands`
runs multiple sub tests. This means, when one of the sub tests fails,
the failed container logs are printed at the end for the parent test,
which makes it harder to aggregate properly in Jenkins. Furthermore,
when multiple sub tests fail, you still only print the logs once at the
end after all tests ran.

This commit suppresses this behavior and instead prints out all the
container logs for every failed sub test. This will guarantee the logs
for the failed tests are grouped with the test, and it will allow us to
distinguish the logs between failed tests.

To accomplish, a new method was added on the suite to reuse the internal
`printFailedContainerLogs` function for the suite.