feat(zbctl): add hidden maxJobsHandle flags
Adds a new `maxJobsHandle` flag to the `createWorker` command, allowing
a worker to exit early if it processed a specific number of jobs. When
set to 0 (default), the worker will process an unlimited number of jobs
and never exit.

This is especially useful in testing, where it's difficult to figure out
if we've observed the expected behavior, as the default worker just
hangs forever. We can make this flag public later if it seems
interesting for users.